refactor(livekit): split 1450-line file into per-concern modules
No behavior change. Replaces commands/livekit.rs with commands/livekit/
containing one file per concern:

  mod.rs           top-level docs, re-exports, dispatch_data helper
  jwt.rs           LiveKitClaims, VideoGrants, make_token, make_view_token
  admin_api.rs     AdminGrants/Claims, make_admin_token, twirp_base,
                   RsParticipantsResp, room_service_list_participants
  identity.rs      user_id_from_identity, lookup_avatar_url,
                   lookup_avatar_url_for_identity,
                   enrich_participants_with_avatars
  participants.rs  VoiceParticipantInfo, VoiceRoomCount,
                   list_voice_room_counts, list_voice_participants
  publish.rs       publish_to_user_inbox/room_server,
                   publish_new/edited/deleted_message_to_room,
                   publish_membership/member_role_changed_to_room,
                   publish_voice_presence, publish_typing, publish_ping,
                   start_call, cancel_call, StartCallResult
  realtime.rs      subscribe_realtime, connect_rooms, emit_presence,
                   emit_room_initial_presence
  legacy.rs        get_livekit_token, get_livekit_view_token, get_livekit_url

Public surface at pollis_core::commands::livekit::* is unchanged. Tests:
184 unit + 26 integration flows pass, same as baseline.
